requirements.txt
Package              Version
-------------------- -----------
amqp                 5.3.1
asgiref              3.10.0
billiard             4.2.2
celery               5.4.0
certifi              2025.10.5
cfgv                 3.4.0
charset-normalizer   3.4.4
click                8.3.0
click-didyoumean     0.3.1
click-plugins        1.1.1.2
click-repl           0.3.0
colorama             0.4.6
crispy-bootstrap5    2024.10
distlib              0.4.0
Django               5.2.6
django-allauth       65.4.1
django-appconf       1.1.0
django-celery-email  3.0.0
django-cors-headers  4.7.0
django-crispy-forms  2.4
django-debug-toolbar 5.0.1
django-filter        24.3
django-imagekit      5.0.0
django-jazzmin       3.0.1
django-js-asset      3.1.2
django-mptt          0.16.0
django-redis         5.4.0
django-tinymce       4.1.0
djangorestframework  3.16.1
drf-yasg             1.21.8
filelock             3.20.0
filetype             1.2.0
identify             2.6.15
idna                 3.11
inflection           0.5.1
isodate              0.7.2
kombu                5.5.4
lxml                 6.0.2
m3u8                 6.0.0
Markdown             3.7
nodeenv              1.9.1
packaging            25.0
pilkit               3.0
pillow               11.1.0
pip                  25.3
platformdirs         4.5.0
pre_commit           4.1.0
prompt_toolkit       3.0.52
psutil               6.1.1
psycopg              3.2.4
psycopg-binary       3.2.4
psycopg-pool         3.2.7
pysubs2              1.8.0
python-dateutil      2.9.0.post0
python-dotenv        1.2.1
python3-saml         1.16.0
pytz                 2025.2
PyYAML               6.0.3
redis                7.0.1
requests             2.32.3
sentry-sdk           2.23.1
six                  1.17.0
sqlparse             0.5.3
typing_extensions    4.15.0
tzdata               2025.2
uritemplate          4.2.0
urllib3              2.5.0
vine                 5.1.0
virtualenv           20.35.4
wcwidth              0.2.14
xmlsec               1.3.16
